% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_assignments}
\alias{detect_assignments}
\title{Detect variable assignments in R code (supports <-, =, ->, <<-, assign())}
\usage{
detect_assignments(code_text)
}
\description{
Detect variable assignments in R code (supports <-, =, ->, <<-, assign())
}
\keyword{internal}
