% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{disable_autoload}
\alias{disable_autoload}
\title{Disable Auto-load on RStudio Startup}
\usage{
disable_autoload()
}
\description{
Removes library(rstudiomcp) from project .Rprofile and disables auto-start
}
